requirements.txt
coreapi==2.0.9
defusedxml==0.4.1
Django==1.10
django-braces==1.10.0
django-crispy-forms==1.6.1
django-filter==1.0.0
django-guardian==1.4.6
django-oauth-toolkit==0.10.0
django-uwsgi==0.1.6
djangorestframework==3.5.3
djangorestframework-oauth==1.1.0
djangorestframework-xml==1.3.0
itypes==1.1.0
jedi==0.9.0
Markdown==2.6.7
mysqlclient==1.3.9
oauthlib==1.0.3
PyMySQL==0.7.9
pytz==2016.7
requests==2.12.1
git+git://github.com/RockyLiys/pip_test.git
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
uWSGI==2.0.14
uwsgitop==0.9
